Move load_results_manually to test_enkf_main
